apdate route rechnolog

+++ backend/db/migrations/20220707142213-create-production.js
@@ -30,7 +30,7 @@ module.exports = {
       },
       out_amount: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL,
       },
       createdAt: {
         allowNull: false,
